@@ -1,13 +1,21 @@
 module github.com/teamwork/kommentaar
 
-go 1.12
+go 1.20
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/imdario/mergo v0.3.13
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/teamwork/test v0.0.0-20181126061546-2ff8918eb6a4
-	github.com/teamwork/utils v0.0.0-20190114034940-d6a1f27ce92c
+	github.com/teamwork/test v0.0.0-20200108114543-02621bae84ad
+	github.com/teamwork/utils/v2 v2.0.1
+	golang.org/x/text v0.9.0
+	golang.org/x/tools v0.8.0
 	gopkg.in/yaml.v3 v3.0.1
 	zgo.at/sconfig v1.2.2-0.20211017232425-870f818a71b7
+)
+
+require (
+	github.com/Strum355/go-difflib v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/teamwork/utils v1.0.1-0.20230426101410-71bb0b003654 // indirect
+	golang.org/x/mod v0.10.0 // indirect
+	golang.org/x/sys v0.7.0 // indirect
 )
